feat: Enhance camera system with projection type retrieval and utilit… (#8)
* feat: Enhance camera system with projection type retrieval and utility function updates

- Added `getCameraType` method to `ICamera` interface, implemented in `PerspectiveCamera` and `OrthographicCamera` classes to distinguish between projection types.
- Updated documentation for camera utility functions to clarify the role of the graphics API parameter.
- Refactored `screenToWorldRay` and `unproject` functions for improved readability.
- Adjusted `build_linux.sh` permissions for execution.
- Added unit tests to validate camera type retrieval functionality.

* fix: Refactor distance checks in PerspectiveCamera for improved readability

- Replaced hardcoded epsilon values with a constant `kEpsilon` for distance comparisons in `rotateAroundTarget` method.
- Enhanced clarity of rotation logic by separating yaw and pitch operations, ensuring up vector remains orthogonal to direction.

+++ tests/camera_utils_test.cpp
@@ -142,8 +142,8 @@ TEST(CameraUtils, FitToAABB_ProjectsInsideNDC) {
     const float ndc_max = 1.0f + ndc_tolerance;
 
     // Project all 8 AABB corners and verify each falls within NDC [-1, 1] x [-1, 1]
-    for (unsigned int i = 0; i < 8; ++i) {
-        Vec3f corner = aabb.corner(static_cast<uint32_t>(i));
+    for (uint32_t i = 0; i < 8u; ++i) {
+        Vec3f corner = aabb.corner(i);
         Vec3f screen = project(*cam, corner, viewport_w, viewport_h);
         float ndc_x = (screen.x() / viewport_w) * 2.0f - 1.0f;
         float ndc_y = (screen.y() / viewport_h) * 2.0f - 1.0f;
